+++ R/example_dataset.R
@@ -1,12 +1,11 @@
 #' Example of dataset
 #'
 #' Toy example to launch functions.
-#' The dimensions of the dataset are 100 rows and 50 columns and this dataset contains 50 spatial-time series.
+#' @format The dimensions of the dataset are 20 rows and 12 columns and this dataset contains 12 spatial-time series.
 #'
-#' @docType data
 #'
 #' @usage example_dataset
 #'
-#'
 #' @keywords example_dataset
+#'
 "example_dataset"