+++ NEWS.md
@@ -1,3 +1,10 @@
+# STMotif 2.0.0
+
+* New example dataset.
+* Fix some functions.
+* New visualization functions.
+
+
 # STMotif 1.0.1
 
 * Fix some functions.
@@ -16,3 +23,5 @@
 # STMotif 0.1.0
 
 * Fist version of the STMotif package.
+
+